+++ client_chargepoints_list.go
@@ -1,11 +1,7 @@
 package monta
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
-	"fmt"
-	"net/http"
 	"net/url"
 	"strconv"
 )
@@ -33,21 +29,8 @@ func (c *Client) ListChargePoints(
 	ctx context.Context,
 	request *ListChargePointsRequest,
 ) (_ *ListChargePointsResponse, err error) {
-	const method, path = http.MethodGet, "/v1/charge-points"
-	defer func() {
-		if err != nil {
-			err = fmt.Errorf("%s %s: %w", method, path, err)
-		}
-	}()
-	var requestBody bytes.Buffer
-	if err := json.NewEncoder(&requestBody).Encode(&request); err != nil {
-		return nil, err
-	}
-	requestURL, err := url.Parse(apiHost + path)
-	if err != nil {
-		return nil, err
-	}
-	query := requestURL.Query()
+	path := "/v1/charge-points"
+	query := url.Values{}
 	if request.Page > 0 {
 		query.Set("page", strconv.Itoa(request.Page))
 	}
@@ -57,27 +40,5 @@ func (c *Client) ListChargePoints(
 	if request.SiteID != nil {
 		query.Set("siteId", strconv.Itoa(int(*request.SiteID)))
 	}
-	requestURL.RawQuery = query.Encode()
-	httpRequest, err := http.NewRequestWithContext(ctx, method, requestURL.String(), &requestBody)
-	if err != nil {
-		return nil, err
-	}
-	if err := c.setAuthorization(ctx, httpRequest); err != nil {
-		return nil, err
-	}
-	httpResponse, err := c.httpClient.Do(httpRequest)
-	if err != nil {
-		return nil, err
-	}
-	defer func() {
-		_ = httpResponse.Body.Close()
-	}()
-	if httpResponse.StatusCode != http.StatusOK {
-		return nil, newStatusError(httpResponse)
-	}
-	var response ListChargePointsResponse
-	if err := json.NewDecoder(httpResponse.Body).Decode(&response); err != nil {
-		return nil, err
-	}
-	return &response, nil
+	return listEntity[ListChargePointsResponse](ctx, c, path, query)
 }
