+++ client_charges_list.go
@@ -1,11 +1,7 @@
 package monta
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
-	"fmt"
-	"net/http"
 	"net/url"
 	"strconv"
 )
@@ -31,25 +27,9 @@ type ListChargesResponse struct {
 }
 
 // ListCharges to retrieve your charge points.
-func (c *Client) ListCharges(
-	ctx context.Context,
-	request *ListChargesRequest,
-) (_ *ListChargesResponse, err error) {
-	const method, path = http.MethodGet, "/v1/charges"
-	defer func() {
-		if err != nil {
-			err = fmt.Errorf("%s %s: %w", method, path, err)
-		}
-	}()
-	var requestBody bytes.Buffer
-	if err := json.NewEncoder(&requestBody).Encode(&request); err != nil {
-		return nil, err
-	}
-	requestURL, err := url.Parse(apiHost + path)
-	if err != nil {
-		return nil, err
-	}
-	query := requestURL.Query()
+func (c *Client) ListCharges(ctx context.Context, request *ListChargesRequest) (*ListChargesResponse, error) {
+	path := "/v1/charges"
+	query := url.Values{}
 	if request.Page > 0 {
 		query.Set("page", strconv.Itoa(request.Page))
 	}
@@ -62,27 +42,5 @@ func (c *Client) ListCharges(
 	if request.ChargePointID != nil {
 		query.Set("chargePointId", strconv.Itoa(int(*request.ChargePointID)))
 	}
-	requestURL.RawQuery = query.Encode()
-	httpRequest, err := http.NewRequestWithContext(ctx, method, requestURL.String(), &requestBody)
-	if err != nil {
-		return nil, err
-	}
-	if err := c.setAuthorization(ctx, httpRequest); err != nil {
-		return nil, err
-	}
-	httpResponse, err := c.httpClient.Do(httpRequest)
-	if err != nil {
-		return nil, err
-	}
-	defer func() {
-		_ = httpResponse.Body.Close()
-	}()
-	if httpResponse.StatusCode != http.StatusOK {
-		return nil, newStatusError(httpResponse)
-	}
-	var response ListChargesResponse
-	if err := json.NewDecoder(httpResponse.Body).Decode(&response); err != nil {
-		return nil, err
-	}
-	return &response, nil
+	return listEntity[ListChargesResponse](ctx, c, path, query)
 }
