+++ client_sites_list.go
@@ -1,11 +1,7 @@
 package monta
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
-	"fmt"
-	"net/http"
 	"net/url"
 	"strconv"
 )
@@ -27,49 +23,14 @@ type ListSitesResponse struct {
 }
 
 // ListSites to retrieve your (charge) sites.
-func (c *Client) ListSites(ctx context.Context, request *ListSitesRequest) (_ *ListSitesResponse, err error) {
-	const method, path = http.MethodGet, "/v1/sites"
-	defer func() {
-		if err != nil {
-			err = fmt.Errorf("%s %s: %w", method, path, err)
-		}
-	}()
-	var requestBody bytes.Buffer
-	if err := json.NewEncoder(&requestBody).Encode(&request); err != nil {
-		return nil, err
-	}
-	requestURL, err := url.Parse(apiHost + path)
-	if err != nil {
-		return nil, err
-	}
-	query := requestURL.Query()
+func (c *Client) ListSites(ctx context.Context, request *ListSitesRequest) (*ListSitesResponse, error) {
+	path := "/v1/sites"
+	query := url.Values{}
 	if request.Page > 0 {
 		query.Set("page", strconv.Itoa(request.Page))
 	}
 	if request.PerPage > 0 {
 		query.Set("perPage", strconv.Itoa(request.PerPage))
 	}
-	requestURL.RawQuery = query.Encode()
-	httpRequest, err := http.NewRequestWithContext(ctx, method, requestURL.String(), &requestBody)
-	if err != nil {
-		return nil, err
-	}
-	if err := c.setAuthorization(ctx, httpRequest); err != nil {
-		return nil, err
-	}
-	httpResponse, err := c.httpClient.Do(httpRequest)
-	if err != nil {
-		return nil, err
-	}
-	defer func() {
-		_ = httpResponse.Body.Close()
-	}()
-	if httpResponse.StatusCode != http.StatusOK {
-		return nil, newStatusError(httpResponse)
-	}
-	var response ListSitesResponse
-	if err := json.NewDecoder(httpResponse.Body).Decode(&response); err != nil {
-		return nil, err
-	}
-	return &response, nil
+	return listEntity[ListSitesResponse](ctx, c, path, query)
 }
